Extract string unescaping to a module
